+++ src/client/components/Buttons/index.js
@@ -4,8 +4,29 @@
  *************************************
  */
 
-import React, { Fragment } from 'react';
+import React, { Component, Fragment } from 'react';
 
+
+/*-- Apply Third-party plugins --*/
+//1) Bootstrap
+import '../_third-party-plugins/Bootstrap/scss/bootstrap.scss';  
+
+
+//2) Fontawesome
+import '../_third-party-plugins/FontAwesome/scss/fontawesome.scss';  
+import '../_third-party-plugins/FontAwesome/scss/v4-shims.scss';  
+import '../_third-party-plugins/FontAwesome/scss/solid.scss';  
+import '../_third-party-plugins/FontAwesome/scss/brands.scss';  
+import '../_third-party-plugins/FontAwesome/scss/regular.scss';  
+
+
+
+/*-- Apply global scripts and styles --*/
+import '../_utilities/scss/_all.scss';
+import '../_utilities/scss/rtl/_all-rtl.scss';
+import { __ } from '../_utilities/js/_all.js';
+
+/*-- Apply this component styles --*/
 import './styles/index.scss';
 
 
@@ -30,15 +51,15 @@ const styles = {
 
 
 
-export default class Button extends React.Component {
+export default class Button extends Component {
 	constructor(props) {
 		super(props);
 	}
 	render() {
 		return (
 		  <Fragment>
-			  <button className='ukr-button' type='button' style={styles[this.props.bgColor] || styles['info']}>
-				{this.props.btnName || 'Default'}
+			  <button className='ukr-button' type='button' style={styles[this.props.ukrBtnBgColor] || styles['info']}>
+				{this.props.ukrBtnName || 'Default'}
 			  </button>
 		  </Fragment>
 		)