+++ src/client/components/_utilities/js/_all.js
@@ -1,0 +1,15 @@
+
+/**
+
+	TABLE OF CONTENTS FOR `ALL UTILITIES`
+	---------------------------
+
+
+
+*/
+import __ from "./dom.js";
+
+
+export { 
+	__ 
+};