+++ src/client/views/_pages/Home/index.js
@@ -1,9 +1,10 @@
 import React, { Fragment } from 'react';
 import Button from '../../../components/Buttons/index.js';
+import Footer from '../../../components/Footer/index.js';
 
 let spreadOperator = {
-  bgColor: 'info',
-  btnName: 'Spread Operator',
+  ukrBtnBgColor: 'info',
+  ukrBtnName: 'Spread Operator',
 };
 
 
@@ -12,18 +13,44 @@ export default ({ staticContext = {} }) => {
     
     return (
 	  <Fragment>
+		    
+		    <div className="uix-header__placeholder js-uix-header__placeholder-autoheight"></div>
+		   
+            <main id="uix-maincontent">
 		
-            <div className="content">
-                <h1>Home!</h1>
+				{/*
+				<!-- Content   
+				====================================================== -->	
+				*/}
+				<section className="uix-spacing--s">
+					<div className="container">
+							<div className="row">
+								<div className="col-12">
 		
-		        <h3>Buttons:</h3>
-				<Button bgColor='' btnName=''/>
-				<Button bgColor='success' btnName='success'/>
-				<Button bgColor='info' btnName='info'/>
-				<Button bgColor='danger' btnName='danger'/>
-				<Button bgColor='warning' btnName='warning'/>
-				<Button {...spreadOperator}/>
-            </div>
+									<h1>Home!</h1>
+
+									<h3>Buttons:</h3>
+									<Button ukrBtnBgColor='' ukrBtnName=''/>
+									<Button ukrBtnBgColor='success' ukrBtnName='success'/>
+									<Button ukrBtnBgColor='info' ukrBtnName='info'/>
+									<Button ukrBtnBgColor='danger' ukrBtnName='danger'/>
+									<Button ukrBtnBgColor='warning' ukrBtnName='warning'/>
+									<Button {...spreadOperator}/>
+
+								</div>
+							</div>
+							{/*<!-- .row end -->*/}
+
+
+					</div>
+					{/*<!-- .container end -->*/}
+				</section>
+		
+		
+
+            </main>
+		
+		    <Footer />
        
         
       </Fragment>