+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 
 
+#### = 0.0.15beta (April 7, 2021) =
+
+* Standardize the navigation menu of the presentation.
+* Add independent component demos which embeds Bootstrap, videojs, tweenmax and fontawesome.
+
 
 #### = 0.0.1beta (April 6, 2021) =